requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
dj-rest-auth==6.0.0
Django==5.1.2
django-allauth==65.1.0
django-cors-headers==4.5.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
drf-spectacular==0.27.2
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.1
pillow==11.0.0
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.3
